Change artifact name, package name for readability, remove unused files
