+++ services/callback/Dockerfile
@@ -12,7 +12,7 @@ COPY ./services/pom.xml ${WORK_DIR}/services/pom.xml
 COPY ./services/distribution/pom.xml ${WORK_DIR}/services/distribution/pom.xml
 COPY ./services/submission/pom.xml ${WORK_DIR}/services/submission/pom.xml
 COPY ./services/callback/pom.xml ${WORK_DIR}/services/callback/pom.xml
-COPY ./services/federationdownload/pom.xml ${WORK_DIR}/services/federationdownload/pom.xml
+COPY ./services/federation-download/pom.xml ${WORK_DIR}/services/federation-download/pom.xml
 
 WORKDIR ${WORK_DIR}
