+++ services/federation-download/Dockerfile
@@ -22,16 +22,16 @@ RUN ${WORK_DIR}/mvnw dependency:go-offline --file ${WORK_DIR}/pom.xml -B
 # copy rest of files
 COPY . ${WORK_DIR}/
 
-RUN mkdir -p /usr/sap/submission-service
+RUN mkdir -p /usr/sap/federation-download-service
 # hadolint ignore=SC2086
 RUN ${WORK_DIR}/mvnw --batch-mode -DskipTests=true --file ${WORK_DIR}/pom.xml ${MAVEN_ARGS} clean install
-RUN cp ${WORK_DIR}/services/submission/target/submission-*.jar /usr/sap/submission-service/submission.jar
+RUN cp ${WORK_DIR}/services/federation-download/target/federation-download-*.jar /usr/sap/federation-download-service/federation-download.jar
 RUN cp ${WORK_DIR}/scripts/DpkgHelper.java /DpkgHelper.java
 
 FROM gcr.io/distroless/java:11
 COPY --from=build /DpkgHelper.java .
-COPY --from=build /usr/sap/submission-service/submission.jar .
+COPY --from=build /usr/sap/federation-download-service/federation-download.jar .
 RUN ["java", "DpkgHelper.java"]
 USER 65534:65534
 ENTRYPOINT ["/usr/bin/java", "-Djdk.tls.namedGroups=\"secp256r1,secp384r1,brainpoolP256r1,brainpoolP384r1,brainpoolP512r1,ffdhe2048,ffdhe3072,ffdhe4096\"", "-jar"]
-CMD ["submission.jar"]
+CMD ["federation-download.jar"]
