+++ services/federation-download/src/main/resources/application-disable-ssl-server.yaml
