+++ services/federation-download/src/main/resources/bootstrap-cloud.yaml
