+++ services/federation-download/src/main/resources/bootstrap.yaml
