--- services/federationdownload/src/main/resources/application.yaml
@@ -1,66 +1,0 @@
----
-logging:
-  level:
-    org:
-      springframework:
-        web: INFO
-    app:
-      coronawarn: INFO
-
-services:
-  federationdownload:
-    # The ApacheHttpClient's connection pool size.
-    connection-pool-size: 200
-    federation-gateway:
-      base-url: ${FEDERATION_GATEWAY_BASE_URL:http://localhost:8005}
-      path: /diagnosiskeys/download/
-    # Local paths, that are used during the export creation.
-    paths:
-      # The output path.
-      output: out
-      # The location of the private key.
-      privatekey: ${VAULT_FILESIGNING_SECRET}
-    # Configuration for the exported archive, that is saved on the S3-compatible storage.
-    tek-export:
-      # The TEK file name included in the zip archive, containing the list of diagnosis keys.
-      file-name: export.bin
-      # The TEK file header.
-      file-header: EK Export v1
-      # The fixed (ensured by right whitespace padding) TEK file header width.
-      file-header-width: 16
-    # Configuration for the API which is used by the mobile app to query diagnosis keys.
-    api:
-      version-path: version
-      version-v1: v1
-      country-path: country
-      country-germany: DE
-      date-path: date
-      hour-path: hour
-      diagnosis-keys-path: diagnosis-keys
-      parameters-path: configuration
-      app-config-file-name: app_config
-
-
-
-spring:
-  main:
-    web-application-type: NONE
-  # Postgres configuration
-  flyway:
-    enabled: true
-    locations: classpath:/db/migration, classpath:/db/specific/{vendor}
-    password: ${POSTGRESQL_PASSWORD_FLYWAY:local_setup_flyway}
-    user: ${POSTGRESQL_USER_FLYWAY:local_setup_flyway}
-
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://${POSTGRESQL_SERVICE_HOST}:${POSTGRESQL_SERVICE_PORT}/${POSTGRESQL_DATABASE}?ssl=true&sslmode=verify-full&sslrootcert=${SSL_POSTGRES_CERTIFICATE_PATH}&sslcert=${SSL_DISTRIBUTION_CERTIFICATE_PATH}&sslkey=${SSL_DISTRIBUTION_PRIVATE_KEY_PATH}
-    username: ${POSTGRESQL_USER_DISTRIBUTION:local_setup_download}
-    password: ${POSTGRESQL_PASSWORD_DISTRIBUTION:local_setup_download}
-
-feign:
-  client:
-    config:
-      default:
-        connect-timeout: 5000
-        read-timeout: 5000
