--- services/federationdownload/src/main/resources/master-config/app-config.yaml
@@ -1,17 +1,0 @@
-#----------------------------------------------------------------------
-# This is the Corona Warn App master configuration file.
-# ----------------------------------------------------------------------
-#
-# This configuration file will be fetched by the mobile app in order to function.
-#
-# min-risk-score: exposure incidents with scores lower than the value of this property will be excluded by the ExposureNotification framework.
-#
-# For details regarding other properties - please see their respective configuration files.
-#
-# Change this file with caution!
-
-min-risk-score: 11
-attenuation-duration: !include attenuation-duration.yaml
-risk-score-classes: !include risk-score-classification.yaml
-exposure-config: !include exposure-config.yaml
-app-version: !include app-version-config.yaml
