--- services/federationdownload/src/main/resources/master-config/app-version-config.yaml
@@ -1,25 +1,0 @@
-# This is the Application Version Configuration master file which contains information for clients
-# about the latest and minimum supported mobile app versions on Android and iOS.
-#
-# The latest version must not be lower than the min version.
-#
-# Change this file with caution!
-
-ios:
-  latest:
-    major: 0
-    minor: 8
-    patch: 2
-  min:
-    major: 0
-    minor: 5
-    patch: 0
-android:
-  latest:
-    major: 1
-    minor: 0
-    patch: 4
-  min:
-    major: 1
-    minor: 0
-    patch: 4
