--- services/federationdownload/src/main/resources/master-config/exposure-config.yaml
@@ -1,58 +1,0 @@
-# This is the Exposure Config master file which defines the risk parameters for the mobile clients
-# Change this file with caution.
-#
-# Weights must be in range of 0.001 to 100
-# Parameter Scores must be in range of 0 to 8.
-#
-# Further documentation:
-# https://developer.apple.com/documentation/exposurenotification/enexposureconfiguration
-#
-
-# Weight Section
-
-transmission_weight: 50
-duration_weight: 50
-attenuation_weight: 50
-days_weight: 20
-
-# Parameters Section
-
-transmission:
-  app_defined_1: 1
-  app_defined_2: 2
-  app_defined_3: 3
-  app_defined_4: 4
-  app_defined_5: 5
-  app_defined_6: 6
-  app_defined_7: 7
-  app_defined_8: 8
-
-duration:
-  eq_0_min: 0
-  gt_0_le_5_min: 0
-  gt_5_le_10_min: 0
-  gt_10_le_15_min: 1
-  gt_15_le_20_min: 1
-  gt_20_le_25_min: 1
-  gt_25_le_30_min: 1
-  gt_30_min: 1
-
-days_since_last_exposure:
-  ge_14_days: 5
-  ge_12_lt_14_days: 5
-  ge_10_lt_12_days: 5
-  ge_8_lt_10_days: 5
-  ge_6_lt_8_days: 5
-  ge_4_lt_6_days: 5
-  ge_2_lt_4_days: 5
-  ge_0_lt_2_days: 5
-
-attenuation:
-  gt_73_dbm: 0
-  gt_63_le_73_dbm: 1
-  gt_51_le_63_dbm: 1
-  gt_33_le_51_dbm: 1
-  gt_27_le_33_dbm: 1
-  gt_15_le_27_dbm: 1
-  gt_10_le_15_dbm: 1
-  le_10_dbm: 1
